保存

+++ client/src/main/java/com/game/module/player/PlayerInfo.java
@@ -32,6 +32,6 @@ public class PlayerInfo {
     private int level;
     private int careerId;
     private int golds;
-    private int sceneId;
+    private long sceneId;
     private PlayerBattle playerBattle;
 }
